Remove transactions from Category table.
Change to have Transactions link to the Category table and allow it to be null.
A null category field in the Transaction table means it is the single parent transaction representing the entire summary transaction.
In this case, there will be child transactions that show all the line items that make up the transaction.